prover: insertion_circuit: adjust response for use in contracts
TODO this commit message is not very accurate, please fix

Dummy data we use in world-id-contracts need post root value, 4844
evaluation and 4844 commitment, all as BN254-reduced 256-bit numbers.
Modify the response to contain all these data in that format, so it
is possible to regenerate the same test data with gnark-mbu, simply
by calling commands saved in world-id-contracts tests' comments
as reference.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>